:sparkles: update types and implement calcMedia

+++ src/main/java/com/remedios/lucas/curso/alunoDisciplina/AlunoDisciplina.java
@@ -17,12 +17,12 @@ public class AlunoDisciplina {
     Long idDisciplina;
     String nomeDisciplina;
     String nomeProfessor;
-    Long nota1;
-    Long nota2;
-    Long notaAtividade;
-    Long media;
+     Double nota1;
+     Double nota2;
+    Double notaAtividade;
+    Double media;
 
-    public AlunoDisciplina(Long idAlunoDisciplina, Long idAluno, String nomeAluno, Long idDisciplina, String nomeDisciplina, String nomeProfessor, Long nota1, Long nota2, Long notaAtividade, Long media) {
+    public AlunoDisciplina(Long idAlunoDisciplina, Long idAluno, String nomeAluno, Long idDisciplina, String nomeDisciplina, String nomeProfessor, Double nota1, Double nota2, Double notaAtividade, Double media) {
         this.idAlunoDisciplina = idAlunoDisciplina;
         this.idAluno = idAluno;
         this.nomeAluno = nomeAluno;
@@ -107,35 +107,35 @@ public class AlunoDisciplina {
         this.nomeProfessor = nomeProfessor;
     }
 
-    public Long getNota1() {
+    public Double getNota1() {
         return nota1;
     }
 
-    public void setNota1(Long nota1) {
+    public void setNota1(Double nota1) {
         this.nota1 = nota1;
     }
 
-    public Long getNota2() {
+    public Double getNota2() {
         return nota2;
     }
 
-    public void setNota2(Long nota2) {
+    public void setNota2(Double nota2) {
         this.nota2 = nota2;
     }
 
-    public Long getNotaAtividade() {
+    public Double getNotaAtividade() {
         return notaAtividade;
     }
 
-    public void setNotaAtividade(Long notaAtividade) {
+    public void setNotaAtividade(Double notaAtividade) {
         this.notaAtividade = notaAtividade;
     }
 
-    public Long getMedia() {
+    public Double getMedia() {
         return media;
     }
 
-    public void setMedia(Long media) {
+    public void setMedia(Double media) {
         this.media = media;
     }
 }
